Increase number of allowed registers in register classes to 64k instead of 256. Widen register class ID to 16-bits. Widen register size and alignment to be up to 64k bytes instead of 256 bytes. This partially reverts r152019 to be less restrictive.
llvm-svn: 152100

+++ llvm/include/llvm/MC/MCRegisterInfo.h
@@ -31,10 +31,10 @@ public:
   const char *Name;
   const iterator RegsBegin;
   const uint8_t *const RegSet;
-  const uint8_t RegsSize;
-  const uint8_t RegSetSize;
-  const uint8_t ID;
-  const uint8_t RegSize, Alignment; // Size & Alignment of register in bytes
+  const uint16_t RegsSize;
+  const uint16_t RegSetSize;
+  const uint16_t ID;
+  const uint16_t RegSize, Alignment; // Size & Alignment of register in bytes
   const int8_t CopyCost;
   const bool Allocatable;
